backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
tortoise-orm==0.20.0
aerich==0.7.2
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
aioredis==2.0.1

# WebSocket
websockets==12.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
